Tear down UpgradeLog when the Upgrade is missing
Normally, the UpgradeLog CR will be deleted when the corresponding
Upgrade CR is removed. But this provides us a double insurance to
clean the mess up.

Note: this commit only addresses the situation that happened at the
beginning of the UpgradeLogReady phase. No further checks after the
point in time.

Signed-off-by: Zespre Chang <zespre.chang@suse.com>